algo qui marche enfin

+++ project/src/main/java/org/apache/project/modele/PlanDeVille.java
@@ -34,7 +34,7 @@ public class PlanDeVille extends Observable {
 		return troncons;
 	}
 	
-	public Intersection getIntersectionById(int id) {
-		return intersections.get((long)id);
+	public Intersection getIntersectionById(Long id) {
+		return intersections.get(id);
 	}
 }
